refactor: search 패키지 추가
