+++ android/app/src/main/java/io/coursepick/coursepick/presentation/search/OnSelectListener.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.presentation
+package io.coursepick.coursepick.presentation.search
 
 import io.coursepick.coursepick.domain.Place
