+++ android/app/src/main/java/io/coursepick/coursepick/presentation/search/SearchViewModel.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.presentation
+package io.coursepick.coursepick.presentation.search
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -7,6 +7,7 @@ import androidx.lifecycle.viewModelScope
 import io.coursepick.coursepick.data.DefaultSearchRepository
 import io.coursepick.coursepick.domain.Place
 import io.coursepick.coursepick.domain.SearchRepository
+import io.coursepick.coursepick.presentation.Logger
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
