+++ android/app/src/main/res/layout/activity_search.xml
@@ -7,13 +7,13 @@
 
         <variable
             name="adapter"
-            type="io.coursepick.coursepick.presentation.SearchAdapter" />
+            type="io.coursepick.coursepick.presentation.search.SearchAdapter" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.SearchActivity">
+        tools:context=".presentation.search.SearchActivity">
 
         <androidx.appcompat.widget.SearchView
             android:id="@+id/searchView"
